+++ BitcoinSwiftTests/GetHeadersMessageTests.swift
@@ -57,7 +57,7 @@ class GetHeadersMessageTests: XCTestCase {
         0x33, 0x48, 0x0b, 0xa7, 0x44, 0xd8, 0xf7, 0x24]   // blockHashStop
     let data = NSData(bytes: bytes, length: bytes.count)
     if let getHeadersMessage = GetHeadersMessage.fromData(data) {
-      XCTAssertEqual(getHeadersMessage.protocolVersion, 70001)
+      XCTAssertEqual(getHeadersMessage.protocolVersion, UInt32(70001))
       XCTAssertEqual(getHeadersMessage.blockLocatorHashes, [blockLocatorHash0, blockLocatorHash1])
       XCTAssertNotNil(getHeadersMessage.blockHashStop)
       XCTAssertEqual(getHeadersMessage.blockHashStop!, blockHashStop)
@@ -84,7 +84,7 @@ class GetHeadersMessageTests: XCTestCase {
         0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00]   // blockHashStop
     let data = NSData(bytes: bytes, length: bytes.count)
     if let getHeadersMessage = GetHeadersMessage.fromData(data) {
-      XCTAssertEqual(getHeadersMessage.protocolVersion, 70001)
+      XCTAssertEqual(getHeadersMessage.protocolVersion, UInt32(70001))
       XCTAssertEqual(getHeadersMessage.blockLocatorHashes, [blockLocatorHash0, blockLocatorHash1])
       XCTAssertNil(getHeadersMessage.blockHashStop)
     } else {
